Separate classes.
