+++ src/Collidable.cpp
@@ -1,0 +1,1 @@
+#include "Collidable.hpp"
