+++ src/Collidable.hpp
@@ -1,0 +1,12 @@
+#ifndef COLLIDABLE_H
+#define COLLIDABLE_H
+
+namespace smurf{
+    
+    class Collidable
+    {
+    };
+    
+};
+
+#endif // COLLIDABLE_H
