+++ src/StaticTransformation.cpp
@@ -1,0 +1,20 @@
+#include "StaticTransformation.hpp"
+
+smurf::StaticTransformation::StaticTransformation(smurf::Frame* sourceFrame, smurf::Frame* targetFrame, const Eigen::Affine3d& sourceToTarget): 
+    Transformation(sourceFrame, targetFrame), sourceToTarget(sourceToTarget)
+{
+
+}
+
+smurf::StaticTransformation::StaticTransformation(smurf::Frame* sourceFrame, smurf::Frame* targetFrame, const Eigen::Quaterniond& rotation, const Eigen::Vector3d& translation): 
+    Transformation(sourceFrame, targetFrame)
+{
+    sourceToTarget.setIdentity();
+    sourceToTarget.rotate(rotation);
+    sourceToTarget.translation() = translation;
+}
+
+const Eigen::Affine3d& smurf::StaticTransformation::getTransformation() const
+{
+    return sourceToTarget;
+}
