+++ src/Transformation.cpp
@@ -1,0 +1,6 @@
+#include "Transformation.hpp"
+
+smurf::Transformation::Transformation(smurf::Frame* sourceFrame, smurf::Frame* targetFrame) : name(sourceFrame->getName() + "2" + targetFrame->getName()), sourceFrame(sourceFrame), targetFrame(targetFrame)
+{
+
+}
